sqlite_to_postgres/requirements.txt
asgiref==3.8.1
mccabe==0.7.0
psycopg==3.2.3
psycopg-binary==3.2.3
#psycopg2==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
setuptools==65.5.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
